poit-monitor/requirements.txt
# POIT Monitor - Python Dependencies
# ==================================

# Browser automation (bypasses CAPTCHA)
undetected-chromedriver>=3.5.0
selenium>=4.15.0

# Optional: Playwright for GitHub Actions (backup)
# playwright>=1.40.0

# Supabase client
supabase>=2.0.0

# Email via Resend
resend>=0.7.0

# FastAPI for API endpoints (optional, for local testing)
fastapi>=0.104.0
uvicorn>=0.24.0

# Pydantic for data validation
pydantic>=2.0.0

# Async HTTP client (used by supabase)
httpx>=0.25.0

# Environment management (optional)
python-dotenv>=1.0.0
